added run info
